requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
boto3==1.34.113
botocore==1.34.113
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.13
django-celery-beat==2.6.0
django-timezone-field==6.1.0
et-xmlfile==1.1.0
jmespath==1.0.1
kombu==5.3.7
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
parquet==1.3.1
ply==3.11
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyarrow==16.1.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.4
s3transfer==0.10.1
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
thriftpy2==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
